Credit: Jeremy Brewer
Fixed a bug where FITS headers with NAXIS1 and NAXIS2 values with 
NAXIS=0 (these are technically invalid) would project all points to
0, 0.  To fix this, I modified Fits::AddImageDimensions() to ensure
that NAXIS is always >= 2 regardless of whether NAXIS1 and NAXIS2 are 
present.

Additionally, I found that the CDELT3 and CDELT4 and related keywords 
didn't affect the projection as I previously thought.  I thus removed 
the checks for these keywords as all of my test images project properly 
(including the IRAS images which were the initial reason for this 
check).  I'm unsure of what the intial problem I had was, though I 
suspect it was from the WCSAXES keyword which acts like NAXIS.

Finally, I updated wcs2kml.cc to print out the image boundaries.  This 
is useful because the user can see when our projection engine is 
producing funny results (such as a 0 pixel scale, the bug this CL 
addresses).